A3: Improve implementation and documentation

+++ aufgabe3/util.hpp
@@ -4,16 +4,16 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-size_t factorial(size_t n);
+uint64_t factorial(uint64_t n);
 
-size_t ind(vector<unsigned> const &p);
+uint64_t ind(vector<unsigned> const &p);
 
-vector<unsigned> gamma(vector<unsigned> const &p, size_t i);
+vector<unsigned> gamma(vector<unsigned> const &p, unsigned i);
 
-size_t ind_gamma(vector<unsigned> const &p, size_t i);
+uint64_t ind_gamma(vector<unsigned> const &p, unsigned i);
 
-vector<unsigned> ith_permutation(size_t n, size_t i);
+vector<unsigned> ith_permutation(unsigned n, uint64_t i);
 
-vector<unsigned> reverse_and_eat(vector<unsigned> const &p, size_t i);
+vector<unsigned> reverse_and_eat(vector<unsigned> const &p, unsigned i);
 
 #endif
